+++ graphviz/backend/piping.py
@@ -15,6 +15,7 @@ __all__ = ['pipe', 'pipe_string',
 def pipe(engine: str, format: str, data: bytes,
          renderer: typing.Optional[str] = None,
          formatter: typing.Optional[str] = None,
+         neato_no_op: typing.Union[bool, int, None] = None,
          quiet: bool = False) -> bytes:
     """Return ``data`` (``bytes``) piped through ``engine`` into ``format`` as ``bytes``.
 
@@ -24,6 +25,7 @@ def pipe(engine: str, format: str, data: bytes,
         data: Binary (encoded) DOT source bytes to render.
         renderer: Output renderer (``'cairo'``, ``'gd'``, ...).
         formatter: Output formatter (``'cairo'``, ``'gd'``, ...).
+        neato_no_op: Neato layout engine no-op flag.
         quiet: Suppress ``stderr`` output from the layout subprocess.
 
     Returns:
@@ -49,7 +51,9 @@ def pipe(engine: str, format: str, data: bytes,
         The layout command is started from the current directory.
     """
     cmd = dot_command.command(engine, format,
-                              renderer=renderer, formatter=formatter)
+                              renderer=renderer,
+                              formatter=formatter,
+                              neato_no_op=neato_no_op)
     kwargs = {'input': data}
 
     proc = execute.run_check(cmd, capture_output=True, quiet=quiet, **kwargs)
@@ -60,6 +64,7 @@ def pipe_string(engine: str, format: str, input_string: str, *,
                 encoding: str,
                 renderer: typing.Optional[str] = None,
                 formatter: typing.Optional[str] = None,
+                neato_no_op: typing.Union[bool, int, None] = None,
                 quiet: bool = False) -> str:
     """Return ``input_string`` piped through ``engine`` into ``format`` as string.
 
@@ -70,6 +75,7 @@ def pipe_string(engine: str, format: str, input_string: str, *,
         encoding: Encoding to en/decode subprocess stdin and stdout (required).
         renderer: Output renderer (``'cairo'``, ``'gd'``, ...).
         formatter: Output formatter (``'cairo'``, ``'gd'``, ...).
+        neato_no_op: Neato layout engine no-op flag.
         quiet: Suppress ``stderr`` output from the layout subprocess.
 
     Returns:
@@ -96,7 +102,9 @@ def pipe_string(engine: str, format: str, input_string: str, *,
         The layout command is started from the current directory.
     """
     cmd = dot_command.command(engine, format,
-                              renderer=renderer, formatter=formatter)
+                              renderer=renderer,
+                              formatter=formatter,
+                              neato_no_op=neato_no_op)
     kwargs = {'input': input_string, 'encoding': encoding}
 
     proc = execute.run_check(cmd, capture_output=True, quiet=quiet, **kwargs)
@@ -107,6 +115,7 @@ def pipe_lines(engine: str, format: str, input_lines: typing.Iterator[str], *,
                input_encoding: str,
                renderer: typing.Optional[str] = None,
                formatter: typing.Optional[str] = None,
+               neato_no_op: typing.Union[bool, int, None] = None,
                quiet: bool = False) -> bytes:
     r"""Return ``input_lines`` piped through ``engine`` into ``format`` as ``bytes``.
 
@@ -117,6 +126,7 @@ def pipe_lines(engine: str, format: str, input_lines: typing.Iterator[str], *,
         input_encoding: Encode input_lines for subprocess stdin (required).
         renderer: Output renderer (``'cairo'``, ``'gd'``, ...).
         formatter: Output formatter (``'cairo'``, ``'gd'``, ...).
+        neato_no_op: Neato layout engine no-op flag.
         quiet: Suppress ``stderr`` output from the layout subprocess.
 
     Returns:
@@ -143,7 +153,9 @@ def pipe_lines(engine: str, format: str, input_lines: typing.Iterator[str], *,
         The layout command is started from the current directory.
     """
     cmd = dot_command.command(engine, format,
-                             renderer=renderer, formatter=formatter)
+                              renderer=renderer,
+                              formatter=formatter,
+                              neato_no_op=neato_no_op)
     kwargs = {'input_lines': (line.encode(input_encoding) for line in input_lines)}
 
     proc = execute.run_check(cmd, capture_output=True, quiet=quiet, **kwargs)
@@ -154,6 +166,7 @@ def pipe_lines_string(engine: str, format: str, input_lines: typing.Iterator[str
                       encoding: str,
                       renderer: typing.Optional[str] = None,
                       formatter: typing.Optional[str] = None,
+                      neato_no_op: typing.Union[bool, int, None] = None,
                       quiet: bool = False) -> str:
     r"""Return ``input_lines`` piped through ``engine`` into ``format`` as string.
 
@@ -164,6 +177,7 @@ def pipe_lines_string(engine: str, format: str, input_lines: typing.Iterator[str
         encoding: Encoding to en/decode subprocess stdin and stdout (required).
         renderer: Output renderer (``'cairo'``, ``'gd'``, ...).
         formatter: Output formatter (``'cairo'``, ``'gd'``, ...).
+        neato_no_op: Neato layout engine no-op flag.
         quiet: Suppress ``stderr`` output from the layout subprocess.
 
     Returns:
@@ -190,7 +204,9 @@ def pipe_lines_string(engine: str, format: str, input_lines: typing.Iterator[str
         The layout command is started from the current directory.
     """
     cmd = dot_command.command(engine, format,
-                              renderer=renderer, formatter=formatter)
+                              renderer=renderer,
+                              formatter=formatter,
+                              neato_no_op=neato_no_op)
     kwargs = {'input_lines': input_lines, 'encoding': encoding}
 
     proc = execute.run_check(cmd, capture_output=True, quiet=quiet, **kwargs)