+++ CHANGES.rst
@@ -2,8 +2,8 @@ Changelog
 =========
 
 
-Version 0.20 (in development)
------------------------------
+Version 0.20
+------------
 
 Add keyword-only ``neato_no_op`` argument to ``.render()``, ``.pipe()``,
 and stand-alone ``graphviz.render()`` and ``graphviz.pipe()``.